Only support integer stats for optimized performance.

+++ src/util/statistics.cpp
@@ -1,8 +1,8 @@
 #include "util/statistics.h"
 
+#include <limits.h>
 #include <algorithm>
 #include <stddef.h>
-#include <float.h>
 
 void openxc::util::statistics::initialize(DeltaStatistic* stat) {
     stat->total = 0;
@@ -10,13 +10,13 @@ void openxc::util::statistics::initialize(DeltaStatistic* stat) {
 }
 
 void openxc::util::statistics::initialize(Statistic* stat) {
-    stat->min = FLT_MAX;
-    stat->max = FLT_MIN;
+    stat->min = INT_MAX;
+    stat->max = INT_MIN;
     stat->movingAverage = 0;
     stat->alpha = .1;
 }
 
-void openxc::util::statistics::update(Statistic* stat, float newValue) {
+void openxc::util::statistics::update(Statistic* stat, int newValue) {
     stat->min = std::min(newValue, stat->min);
     stat->max = std::max(newValue, stat->max);
 
@@ -24,8 +24,8 @@ void openxc::util::statistics::update(Statistic* stat, float newValue) {
             (1.0 - stat->alpha) * stat->movingAverage;
 }
 
-void openxc::util::statistics::update(DeltaStatistic* stat, float newValue) {
-    float delta = newValue - stat->total;
+void openxc::util::statistics::update(DeltaStatistic* stat, int newValue) {
+    int delta = newValue - stat->total;
     stat->total = newValue;
     update(&stat->statistic, delta);
 }
@@ -38,18 +38,18 @@ float openxc::util::statistics::exponentialMovingAverage(const DeltaStatistic* s
     return exponentialMovingAverage(&stat->statistic);
 }
 
-float openxc::util::statistics::minimum(const Statistic* stat) {
+int openxc::util::statistics::minimum(const Statistic* stat) {
     return stat->min;
 }
 
-float openxc::util::statistics::minimum(const DeltaStatistic* stat) {
+int openxc::util::statistics::minimum(const DeltaStatistic* stat) {
     return stat->statistic.min;
 }
 
-float openxc::util::statistics::maximum(const Statistic* stat) {
+int openxc::util::statistics::maximum(const Statistic* stat) {
     return stat->max;
 }
 
-float openxc::util::statistics::maximum(const DeltaStatistic* stat) {
+int openxc::util::statistics::maximum(const DeltaStatistic* stat) {
     return stat->statistic.max;
 }
